avoid relying on spread props so much
Svelte's Runes seem to be getting a bit confused with the amount of spreading
going on, and I'm losing reactivity somewhere along the line. There's really
no reason for me to be spread-applying props at so many levels; the objects
can be passed around "whole" until the props are actually needed for rendering.